pyproject.toml
[project]
name = "ai-email-assistant-natural-language-to-email-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.177.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "loguru>=0.7.3",
    "openai>=1.98.0",
    "pydantic[email]>=2.11.7",
    "ruff>=0.12.7",
    "uvicorn[standard]>=0.35.0",
]
